change connection mechanism in InputChannelConnection
send_status_message_.info.index

InputChannelConnection inherits from Connection and used to directly call the inherited connect method.
This patch adds a connect method to InputChannelConnection which forwards the call to Connection and posts a status message to the ComputeBuffer.
It uses fi_getname to get the native address and adds it to the status message.