Update .travis [add comment]
Libfabric Version 1.4 is downloaded and configured. The libfabric-dev package is in the debian-sid (https://packages.debian.org/sid/libfabric-dev) but I cannot added to the package list because it's not listed in the whitelist of travis. A new ticket is opened to add the libfabric (https://github.com/travis-ci/apt-package-whitelist/issues/3665)