Fix compilation on macosx
Add a patch for zeromq to properly use the found gnutls during compilation.
Build zeromq without libsodium and libbsd, use internal zeromq code instead.

Check if OpenSSl and gnutl are installed. On macosx check additionally for
zstd. Use found paths and libraries in the project to properly link on macosx.