Stops doxygen from ignoring rmda and libfabric
Access to the rdma and libfabric libraries is guarded in the code by
preprocessor macro determining whether rdma and libfabric are available.
Doxygen's preprocessor filters these code fragments out since the
corresponding preprocessor macros are not defined in the code, but
rather passed to the compiler by the build process. In over to overcome
this problem one must instruct doxygen to define these as well.