Fixes strange notification placement in build
Previously, the notification about an expected error during building the
msconsumer documentation using doxygen was printed on a seperate line. In
particular when building the documentation in parallel, this placement was
strange. It is still not optimal, as when building in parallel the
error does not appear next to the notification, but likely fixing this
minor inconvenience does not pay of.