update Flesnet spelling akin to "CBM TDR Online Systems - Part I"
For the "CBM TDR Online Systems - Part I", we used the spelling
'Flesnet' instead of the older 'FLESnet' or 'FLESNET' spelling. We
update such occurrences to 'Flesnet'. Now, two spellings remain used:
'Flesnet' and 'flesnet', the latter, for example, to refer to the
command line tool (but not exclusively for that purpose).