Make flesnet's command-line option '-e' optional
The command-line option '-e' did not have a default value, so that a
command to start a timeslice processor was mandatory on the command
line.

This makes the use of flesnet less flexible than possible. Actually,
it is not necessary that flesnet initiates the timeslice processor
itself. This can be done asynchronously and independently if
desirable.

For example, in the future scenario of timeslice forwarding, we may
want to start the forwarder more flexible in job scripts by other
means. So, requiring to do that via the '-e' option feels restrictive
and introduces unnecessary coupling in the tool chain.

Now, we provide a default value for the '-e' option, which make that
option optional on the command line. We use "true" as the default
value, which means that flesnet simply calls the UNIX binary 'true' as
dummy timeslice processor.