tsclient: use bool_switch for flag-like options, improve help screen
Use bool_switch for the flag options -a, -b, --histograms, and -R.
This simplifies and clarifies the help screen and specifying but
diabling them does not help much.

Clarify and provide more details in the descriptions of several
other options.

Add a check that the --histograms option needs the -a option.